Plans now hold collections of declarations
This is an upgrade over hashes. Plans are no longer initialized
with their full mappings.  Instead, plans are built. This is
intended to be an API exposed to a DSL at a later time, a step
that is part of a planned move away from yaml files as the plan
container.